src/main/webapp/app/entities/product-order-detail/delete/product-order-detail-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { provideHttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { ProductOrderDetailService } from '../service/product-order-detail.service';

import { ProductOrderDetailDeleteDialogComponent } from './product-order-detail-delete-dialog.component';

describe('ProductOrderDetail Management Delete Component', () => {
  let comp: ProductOrderDetailDeleteDialogComponent;
  let fixture: ComponentFixture<ProductOrderDetailDeleteDialogComponent>;
  let service: ProductOrderDetailService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [ProductOrderDetailDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(ProductOrderDetailDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(ProductOrderDetailDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(ProductOrderDetailService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
